revert cortex-a15 to cortex-a9
this probably sounds like a regression but it's not.  there's no support
in GCC ARM for a variety of cflag optimizations for this arch.  cortex-a15
on the N10 is a transitional chipset.. so lets treat it as such and revert
to a higher level of optimizations until there is more support.

Change-Id: I6fe11d069c2b342d5a2f75f87a7b48298e605452